A number of fixes

+++ scripts/build.sh
@@ -7,5 +7,8 @@ bundle install
 bundle exec rake report:init
 bundle exec rake report:publish
 ebook-convert output/doc.epub output/doc.mobi
-rm .gitignore
-mv output/doc.html index.html
+
+if [ "$1" == "CI" ];then
+  rm .gitignore
+  mv output/doc.html index.html
+fi
